design: 애니메이션 취소 기능 삭제, 기타 자잘한 UI 수정
