+++ src/pages/gacha/components/GachaDropRateInfo.tsx
@@ -8,13 +8,13 @@ const RARITY_INFO: RarityInfo[] = [
   { rarity: 'S', dropRate: '1', color: '#f74600' },
   { rarity: 'A', dropRate: '4', color: '#ffc337' },
   { rarity: 'B', dropRate: '30', color: '#7abf16' },
-  { rarity: 'C', dropRate: '45', color: '#006bff' },
+  { rarity: 'C', dropRate: '45', color: '#108df1' },
   { rarity: 'D', dropRate: '20', color: '#a663ee' },
 ];
 
 export const GachaDropRateInfo = () => (
   <div className='w-48 rounded-lg bg-white/10 px-6 py-4 shadow-md backdrop-blur-sm'>
-    <div className='mb-2 text-center text-2xl font-bold text-white'>
+    <div className='mb-2 text-center text-xl font-bold text-white'>
       획득 확률
     </div>
     <hr className='mb-4' />
