+++ src/pages/gacha/components/PointDisplay.tsx
@@ -7,13 +7,13 @@ interface PointDisplayProps {
 export const PointDisplay = ({ point }: PointDisplayProps) => {
   return (
     <div className='w-48 rounded-lg bg-white/10 px-6 py-4 shadow-md backdrop-blur-sm'>
-      <h2 className='mb-2 text-center text-xl font-bold text-white'>
-        내 포인트
-      </h2>
+      <div className='mb-2 text-center text-xl font-bold text-white'>
+        내 냥코인
+      </div>
       <hr className='mb-4' />
-      <div className='flex text-lg text-white'>
-        <img src={coinImg} alt='coin' className='w-8' />
-        {point.toLocaleString()}
+      <div className='flex justify-around text-lg text-white'>
+        <img src={coinImg} alt='coin' className='w-7' />
+        <p>{point.toLocaleString()}냥</p>
       </div>
     </div>
   );
