+++ src/pages/gacha/components/SummaryItem.tsx
@@ -1,8 +1,20 @@
 import { memo } from 'react';
 
 import { CAT_MAPPINGS, RARITY_TO_IMAGE } from '../constants/catMappings';
-import { RARITY_INFO } from '../constants/rarityInfo';
 import { SummaryItemProps } from '../hooks/usePreloader';
+
+interface RarityInfo {
+  dropRate: string;
+  color: string;
+}
+
+const RARITY_INFO: Record<string, RarityInfo> = {
+  S: { dropRate: '1', color: '#f74600' },
+  A: { dropRate: '4', color: '#ffc337' },
+  B: { dropRate: '30', color: '#7abf16' },
+  C: { dropRate: '45', color: '#108df1' },
+  D: { dropRate: '20', color: '#a663ee' },
+};
 
 export const SummaryItem = memo(({ result }: SummaryItemProps) => {
   const capsuleImages =
@@ -28,7 +40,7 @@ export const SummaryItem = memo(({ result }: SummaryItemProps) => {
         >
           {result.rarity}
         </div>
-        <div className='text-center text-xl font-bold'>
+        <div className='text-center text-xl font-bold text-white'>
           {CAT_MAPPINGS[result.name]?.korName}
         </div>
       </div>
